requirements.txt
flet
psycopg2-binary
sqlalchemy